Refactor upload mechanics into mixin, keep component as presentation logic
